+++ core/src/main/java/com/hortonworks/iotas/service/CatalogService.java
@@ -25,6 +25,7 @@ import com.hortonworks.iotas.topology.TopologyActions;
 import com.hortonworks.iotas.topology.TopologyComponent;
 import com.hortonworks.iotas.topology.TopologyLayoutConstants;
 import com.hortonworks.iotas.topology.TopologyLayoutValidator;
+import com.hortonworks.iotas.topology.TopologyMetrics;
 import com.hortonworks.iotas.util.CoreUtils;
 import com.hortonworks.iotas.util.JarStorage;
 import com.hortonworks.iotas.util.JsonSchemaValidator;
@@ -62,14 +63,13 @@ public class CatalogService {
     private static final String CLUSTER_NAMESPACE = new Cluster().getNameSpace();
     private static final String COMPONENT_NAMESPACE = new Component().getNameSpace();
     private static final String NOTIFIER_INFO_NAMESPACE = new NotifierInfo().getNameSpace();
-    private static final String TOPOLOGY_NAMESPACE = new Topology()
-            .getNameSpace();
+    private static final String TOPOLOGY_NAMESPACE = new Topology().getNameSpace();
     private static final String TAG_NAMESPACE = new Tag().getNameSpace();
     private static final String TAG_STORABLE_MAPPING_NAMESPACE = new TagStorableMapping().getNameSpace();
 
-
     private StorageManager dao;
     private TopologyActions topologyActions;
+    private TopologyMetrics topologyMetrics;
     private JarStorage jarStorage;
     private TagService tagService;
 
@@ -118,9 +118,10 @@ public class CatalogService {
         }
     }
 
-    public CatalogService(StorageManager dao, TopologyActions topologyActions, JarStorage jarStorage) {
+    public CatalogService(StorageManager dao, TopologyActions topologyActions, TopologyMetrics topologyMetrics, JarStorage jarStorage) {
         this.dao = dao;
         this.topologyActions = topologyActions;
+        this.topologyMetrics = topologyMetrics;
         this.jarStorage = jarStorage;
         this.tagService = new CatalogTagService(dao);
     }
@@ -543,6 +544,10 @@ public class CatalogService {
 
     public TopologyActions.Status topologyStatus (Topology topology) throws Exception {
         return this.topologyActions.status(topology);
+    }
+
+    public Map<String, TopologyMetrics.ComponentMetric> getTopologyMetrics (Topology topology) throws Exception {
+        return this.topologyMetrics.getMetricsForTopology(topology);
     }
 
     public Collection<TopologyComponent.TopologyComponentType> listTopologyComponentTypes () {